[Refactor] 패키지 네이밍 변경 #16 - domain -> core - domain/api -> core/presentation - domain/entity -> core/domain
