+++ src/main/java/com/wooringpang/userservice/config/TokenProvider.java
@@ -1,7 +1,7 @@
 package com.wooringpang.userservice.config;
 
-import com.wooringpang.userservice.domain.user.entity.User;
-import com.wooringpang.userservice.domain.user.service.UserService;
+import com.wooringpang.userservice.core.user.domain.User;
+import com.wooringpang.userservice.core.user.service.UserService;
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.SignatureAlgorithm;
