+++ src/main/java/com/wooringpang/userservice/core/log/repository/LoginLogRepository.java
@@ -1,0 +1,7 @@
+package com.wooringpang.userservice.core.log.repository;
+
+import com.wooringpang.userservice.core.log.domain.LoginLog;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface LoginLogRepository extends JpaRepository<LoginLog, Long> {
+}
