+++ src/main/java/com/wooringpang/userservice/core/user/presentation/response/SaveUserResponse.java
@@ -1,0 +1,14 @@
+package com.wooringpang.userservice.core.user.presentation.response;
+
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+public class SaveUserResponse {
+
+    private Long userId;
+}
