+++ src/main/java/com/wooringpang/userservice/core/user/service/param/SaveUserParam.java
@@ -1,0 +1,36 @@
+package com.wooringpang.userservice.core.user.service.param;
+
+import com.wooringpang.userservice.core.user.domain.Role;
+import com.wooringpang.userservice.core.user.domain.User;
+import com.wooringpang.userservice.core.user.domain.UserState;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+
+import java.util.UUID;
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class SaveUserParam {
+
+    private String name;
+    private String email;
+    private String password;
+    private Role role;
+    private UserState userState;
+
+    public User toEntity(BCryptPasswordEncoder passwordEncoder) {
+        return User.createBuilder()
+                .signId(UUID.randomUUID().toString())
+                .name(this.name)
+                .email(this.email)
+                .password(passwordEncoder.encode(this.password))
+                .role(this.role)
+                .userState(this.userState)
+                .build();
+    }
+}
