--- src/main/java/com/wooringpang/userservice/domain/user/api/request/SocialUserResponse.java
@@ -1,22 +1,0 @@
-package com.wooringpang.userservice.domain.user.api.request;
-
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class SocialUserResponse {
-
-    private String id;
-    private String email;
-    private String name;
-
-    @Builder
-    public SocialUserResponse(String id, String email, String name) {
-        this.id = id;
-        this.email = email;
-        this.name = name;
-    }
-}
