--- src/main/java/com/wooringpang/userservice/domain/user/api/response/UpdateUserResponse.java
@@ -1,14 +1,0 @@
-package com.wooringpang.userservice.domain.user.api.response;
-
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-public class UpdateUserResponse {
-
-    private Long userId;
-}
