--- src/main/java/com/wooringpang/userservice/domain/user/dto/UserSearchCondition.java
@@ -1,11 +1,0 @@
-package com.wooringpang.userservice.domain.user.dto;
-
-import com.wooringpang.userservice.common.dto.CommonSearchCondition;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class UserSearchCondition extends CommonSearchCondition {
-}
